+++ packages/jsActions/mobile-resources-native/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 -   We updated the default presentation mode for both "Take Picture" and "Take Picture Advanced" to full screen.
+-   We have added push notification permission request support for Android API level>=33.
 
 ## [5.0.1] Native Mobile Resources - 2023-7-31