+++ packages/jsActions/mobile-resources-native/CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [5.0.2] Native Mobile Resources - 2023-8-22
 ### Fixed
 
 -   We updated the default presentation mode for both "Take Picture" and "Take Picture Advanced" to full screen.
+
+## [2.0.2] ColumnChart
+### Fixed
+
+-   We fixed an issue when running with Mx10 crashes the app.
 
 ## [5.0.1] Native Mobile Resources - 2023-7-31